fix: move from- and toJwk logic to Key class
The reason for doing so is to avoid dependency cycles.
